Use esp_wireguard library both for Arduino and esp-idf frameworks
